Moved hashmap, stack, and tokenlist implementations to internal directory. Adjusted header files.
