+++ src/internal/alloc_stack.c
