+++ src/internal/alloc_stack.h
