+++ src/internal/hash_map.c
