+++ src/internal/hash_map.h
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 #include "../strings/mstring.h"
-#include "object.h"
+#include "../objects/object.h"
 
 typedef Object _value;
 typedef mString _key;
