+++ src/objects/initializers.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "types.h"
+#include "../internal/types.h"
 #include "function.h"
 #include "vector.h"
